% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cet.R
\docType{data}
\name{cet}
\alias{cet}
\title{CET trffic counts and speeds}
\format{A data frame:
\describe{
  \item{highway}{Open Street Map type of street}
  \item{Lanes}{number of lanes from 1 to 10}
  \item{Cars_veh_h}{Number of vehicles per hour}
  \item{Speed_km_h}{Average speed (km/h)}
  \item{Cars_Speed}{Ratio Cars and Speed}
  \item{Trucks_veh_h}{Number of vehicles per hour}
  \item{SpeedT_km_h}{Average speed (km/h)}
  \item{Trucks_Speed}{Ratio Cars and Speed}
}}
\source{
\url{http://www.cetsp.com.br/sobre-a-cet/relatorios-corporativos.aspx}
}
\usage{
data(cet)
}
\description{
#cet <- data.frame(stringsAsFactors=FALSE,
#          street = c("motorway", "motorway", "motorway", "motorway",
#                     "motorway", "motorway", "motorway", "motorway",
#                     "motorway", "motorway", "trunk", "trunk", "trunk", "trunk", "trunk",
#                     "trunk", "trunk", "trunk", "trunk", "trunk", "Primary",
#                     "Primary", "Primary", "Primary", "Primary", "Primary",
#                     "Primary", "Primary", "Primary", "Primary", "Secondary",
#                     "Secondary", "Secondary", "Secondary", "Secondary",
#                     "Secondary", "Secondary", "Secondary", "Secondary",
#                     "Secondary", "Tertiary", "Tertiary", "Tertiary", "Tertiary",
#                     "Tertiary", "Tertiary", "Tertiary", "Tertiary", "Tertiary",
#                     "Tertiary"),
#          lanes = c(10, 9, 8, 7, 6, 5, 4, 3, 2, 1, 10, 9, 8, 7, 6, 5, 4, 3, 2,
#                     1, 10, 9, 8, 7, 6, 5, 4, 3, 2, 1, 10, 9, 8, 7, 6, 5, 4, 3,
#                     2, 1, 10, 9, 8, 7, 6, 5, 4, 3, 2, 1),
#      Cars_veh_h = c(9615.5, 4709.07, 4709.07, 4416.85, 4416.85, 3282.5,
#                     3282.5, 786.5, 786.5, 786.5, 9615.5, 4709.07, 4709.07,
#                     4416.85, 4416.85, 3282.5, 3282.5, 786.5, 786.5, 786.5, 4486.5,
#                     4486.5, 4486.5, 2664, 3233.25, 3311.2, 3311.2, 2697, 2697,
#                     2697, 1671, 1671, 1671, 2388, 2388, 1295, 2527, 626.5, 626.5,
#                     626.5, 1600, 1600, 1600, 1600, 1600, 1600, 1600, 804, 804,
#                     804),
#      Speed_km_h = c(69.15, 60.62, 60.62, 54.01, 54.01, 51.53, 51.53, 16, 16,
#                     16, 69.15, 60.62, 60.62, 54.01, 54.01, 51.53, 51.53, 16,
#                     16, 16, 61.69, 61.69, 61.69, 30.8, 37.3, 35.88, 35.88, 29.5,
#                     29.5, 29.5, 26, 26, 26, 36.64, 36.64, 28.9, 26.43, 27.5,
#                     27.5, 27.5, 22, 22, 22, 22, 22, 22, 22, 15.9, 15.9, 15.9),
#     Cars_Speed = c(139.05, 77.68, 77.68, 81.78, 81.78, 63.7, 63.7, 49.16,
#                     49.16, 49.16, 139.05, 77.68, 77.68, 81.78, 81.78, 63.7,
#                     63.7, 49.16, 49.16, 49.16, 72.73, 72.73, 72.73, 86.49, 86.67,
#                     86.67, 86.67, 91.42, 91.42, 91.42, 64.3, 64.3, 64.3, 65.17,
#                     65.17, 44.81, 95.6, 22.78, 22.78, 22.78, 72.73, 72.73,
#                     72.73, 72.73, 72.73, 72.73, 72.73, 50.57, 50.57, 50.57),
#    Trucks_veh_h = c(101.06, 101.06, 101.06, 101.06, 101.06, 101.06, 101.06,
#                     101.06, 101.06, 101.06, 101.06, 101.06, 101.06, 101.06,
#                     101.06, 101.06, 101.06, 101.06, 101.06, 101.06, 0, 0, 0, 0, 0,
#                     0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
#                     0, 0, 0, 0, 0),
#     SpeedT_km_h = c(50.26, 50.26, 50.26, 50.26, 50.26, 50.26, 50.26, 50.26,
#                     50.26, 50.26, 50.26, 50.26, 50.26, 50.26, 50.26, 50.26,
#                     50.26, 50.26, 50.26, 50.26, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
#                     0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0),
#   Trucks_Speed = c(2.01074413052129, 2.01074413052129, 2.01074413052129,
#                     2.01074413052129, 2.01074413052129, 2.01074413052129,
#                     2.01074413052129, 2.01074413052129, 2.01074413052129,
#                     2.01074413052129, 2.01074413052129, 2.01074413052129,
#                     2.01074413052129, 2.01074413052129, 2.01074413052129, 2.01074413052129,
#                     2.01074413052129, 2.01074413052129, 2.01074413052129,
#                     2.01074413052129, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
#                     0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0)
#)
save(cet, file = "data/cet.rda")
}
\keyword{datasets}
