% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_data.R
\name{fill_data}
\alias{fill_data}
\title{Fill missing data in osm}
\usage{
fill_data(net, what = "lanes", by = "highway", FUN = "mean")
}
\arguments{
\item{what}{Charcter to identify what to aggregate, default "lanes"}

\item{by}{Charcter to identify aggregate by, default "highway". Sometimes can be "fclass".
It is the type of street according to OpenStreetMap covering:
"motorway", "motorway_link", "primary", "primary_link", "secondary",
"secondary_link", "tertiary", "trunk", "trunk_link"}

\item{FUN}{Character to apply function: "sum", "mean", etc.}

\item{df}{Spatial roadnetwork}
}
\description{
\code{\link{fill_data}} Fill missing data in OSM columns
}
\examples{
\dontrun{
a <- clean()
data(osm)
tail(osm)
unique(osm$lanes)
osm$lanes <- fill_data(osm)
unique(osm$lanes)
tail(osm)
}
}
\seealso{
\code{\link{clean}} \code{\link{points_to_streets}} \code{\link{speed_to_flow}}
}
