% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osm.R
\docType{data}
\name{osm}
\alias{osm}
\title{OpenStreetMap roadnetwork for Sao Paulo downtown}
\format{A Spatial Feature :
\describe{
  \item{id}{id of street}
  \item{highway}{type of street}
  \item{maxspeed}{Free speed}
  \item{lanes}{Number of lanes}
}}
\source{
\url{https://www.openstreetmap.org/}
}
\usage{
data(osm)
}
\description{
Spatial Feature
}
\keyword{datasets}
