% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points_to_lines.R
\name{points_to_lines}
\alias{points_to_lines}
\title{From points to lines}
\usage{
points_to_lines(sfpoints, dist, net, verbose = TRUE)
}
\arguments{
\item{sfpoints}{spatial points from \code{\link{clean}}}

\item{dist}{buffer distance to intersect pointts with streets.
IF lat lon, degree, if projected [m]}

\item{net}{Spatial roadnetwork}
}
\description{
\code{\link{points_to_lines}} Aggregate points by street
}
\note{
net must have a column name id to each street.
sfpoints must include names "id", "veh" and "speed".
Use with UTM data (not lat lon)
}
\examples{
\dontrun{
a <- clean()
a
plot(a["type"], axes = T)
library(vein)
data(net)
net$id <- 1:nrow(net@data)
d <- points_to_lines(a, 1/102/47, net)
plot(d[c("MeanSpeed")], axes = T)
plot(d[c("MaxSpeed")], axes = T)
}
}
\seealso{
\code{\link{clean}}
}
