% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speed_to_flow.R
\name{speed_to_flow}
\alias{speed_to_flow}
\title{From points to lines}
\usage{
speed_to_flow(net, speed = "MaxSpeed", highway = "highway",
  lanes = "lanes", method = "methodsx", df_streets)
}
\arguments{
\item{net}{Spatial roadnetwork}

\item{speed}{Charcter to identify column speed}

\item{highway}{Charcter to identify type of street, sometimes called as fclass in osm.}

\item{lanes}{Charcter to identify column lanes}

\item{method}{"methodsx" only in the meantime}

\item{df_street}{optional data.frame with the same dimensions as data(m1),
but values can change. This means that can user internal values or own values.}
}
\description{
\code{\link{points_to_lines}} Aggregate points by street
}
\examples{
\dontrun{
a <- clean()
a
plot(a["type"], axes = T, pch = 16, cex = 0.5, col = "red")
data(osm)
unique(osm$lanes)
osm$lanes <- fill_data(osm)
unique(osm$lanes)
d <- points_to_lines(a, 1/102/47, osm)
#same speed becasue few observations by street
#need more data
plot(d[c("MeanSpeed")], axes = T)
plot(d[c("MaxSpeed")], axes = T)
b <- speed_to_flow(d)
}
}
\seealso{
\code{\link{clean}} \code{\link{points_to_streets}}
}
