% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean}
\alias{clean}
\title{Reads, cleans  and filter GPS data-set}
\usage{
clean(input_file = "data-raw/dados/000000000000.csv", output_file,
  timezone = "Etc/UTC", n = 5, max_speed = 110,
  max_acceleration = 10, verbose = TRUE)
}
\arguments{
\item{input_file}{Character with location of '.csv', or data.table or data.frame
with gps data. The following colnames are required: "veh", "type", "time",
"lon", "lat". "veh" id for each vehicle. "type" type of each vehicle,
"time" character with format "2014-10-10 08:06:40".}

\item{output_file}{Character to store resulting data.table as .csv. If missing
returns data.table}

\item{timezone}{timezone}

\item{n}{Integer, minimum number of consecutive observations by veh}

\item{max_speed}{Numeric, Maximum speed default 110 km/h}

\item{max_acceleration}{Numeric, Maximum acceleration default 10 m/s^2}

\item{verbose}{logical, to show more information}
}
\description{
\code{\link{clean}} Reads, cleans  and filter GPS data-set
}
\examples{
\dontrun{
a <- clean()
}
}
\seealso{
\code{\link{clean}}
}
